進化時のUI調整、スペシャルのダメージテキスト追加、その他プレイ会前の調整

+++ fade.cpp
@@ -124,7 +124,7 @@ void Fade_Draw()
 
 }
 
-void SetFade(float fadeframe, XMFLOAT4 color, FADE_STATE state, SCENE scene)
+void SetFade(int fadeframe, XMFLOAT4 color, FADE_STATE state, SCENE scene)
 {
 	g_Fade.frame = fadeframe;
 	g_Fade.fadecolor = color;
